- Send songrating to client. - Receive songrating as float (like saved in Song). - On RemoteClient destructor check if socket is still connected before calling waitForDisconnect()

+++ src/networkremote/incomingdataparser.h
@@ -48,7 +48,7 @@ signals:
   void SendSongs(const pb::remote::RequestDownloadSongs& request, RemoteClient* client);
   void ResponseSongOffer(RemoteClient* client, bool accepted);
   void SendLibrary(RemoteClient* client);
-  void RateCurrentSong(int);
+  void RateCurrentSong(double);
 
 private:
   Application* app_;
